feat: add types to view/sources
