+++ migration-to-ts/src/components/view/appView.ts
@@ -11,12 +11,12 @@ export class AppView {
         this.sources = new Sources();
     }
 
-    drawNews(data?: EverythingResponse) {
+    drawNews(data: EverythingResponse) {
         const values = data?.articles ? data?.articles : [];
         this.news.draw(values);
     }
 
-    drawSources(data?: SourcesResponse) {
+    drawSources(data: SourcesResponse) {
         const values = data?.sources ? data?.sources : [];
         this.sources.draw(values);
     }
